dom/network/tests/test_networkstats_disabled.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure NetworkStats is not accessible when it is disabled</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();

// Test to ensure NetworkStats is not accessible when it is disabled
SpecialPowers.pushPrefEnv({'set': [["dom.mozNetworkStats.enabled", false]]}, function(){

  ok('mozNetworkStats' in navigator, "navigator.mozNetworkStats should exist");
  is(navigator.mozNetworkStats, null, "mozNetworkStats should be null when not enabled.");

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
